+++ attributes/default.rb
@@ -45,13 +45,46 @@ default['consul']['source_revision'] = 'master'
 
 # Service attributes
 default['consul']['service_mode'] = 'bootstrap'
+# In the cluster mode, set the default cluster size to 3
+default['consul']['bootstrap_expect'] = 3
 default['consul']['data_dir'] = '/var/lib/consul'
 default['consul']['config_dir'] = '/etc/consul.d'
+case node['platform_family']
+when 'debian'
+  default['consul']['etc_config_dir'] = '/etc/default/consul'
+when 'rhel'
+  default['consul']['etc_config_dir'] = '/etc/sysconfig/consul'
+else
+  default['consul']['etc_config_dir'] = '/etc/sysconfig/consul'
+end
+
 default['consul']['servers'] = []
 default['consul']['init_style'] = 'init'   # 'init', 'runit'
 default['consul']['service_user'] = 'consul'
 default['consul']['service_group'] = 'consul'
-default['consul']['gomaxprocs'] = 1
+default['consul']['ports'] = {
+  'dns'      => 8600,
+  'http'     => 8500,
+  'rpc'      => 8400,
+  'serf_lan' => 8301,
+  'serf_wan' => 8302,
+  "server"   => 8300,
+}
+
+# Gossip encryption
+default['consul']['encrypt_enabled'] = false
+default['consul']['encrypt'] = nil
+# TLS support
+default['consul']['verify_incoming'] = false
+default['consul']['verify_outgoing'] = false
+# Cert in pem format
+default['consul']['ca_cert'] = nil
+default['consul']['ca_path'] = "%{config_dir}/ca.pem"
+default['consul']['cert_file'] = nil
+default['consul']['cert_path'] = "%{config_dir}/cert.pem"
+# Cert in pem format. It can be unique for each host
+default['consul']['key_file'] = nil
+default['consul']['key_file_path'] = "%{config_dir}/key.pem"
 
 # Optionally bind to a specific interface
 default['consul']['bind_interface'] = nil
@@ -62,3 +95,4 @@ default['consul']['client_interface'] = nil
 default['consul']['client_addr'] = '0.0.0.0'
 default['consul']['ui_dir'] = '/var/lib/consul/ui'
 default['consul']['serve_ui'] = false
+default['consul']['extra_params'] = {}