+++ spec/fixtures/cookbooks/consul_spec/recipes/service_def_delete.rb
@@ -1,0 +1,5 @@
+include_recipe "consul"
+consul_service_def "dummy" do
+  id "uniqueid"
+  action :delete
+end