+++ spec_helper.rb
@@ -1,4 +1,5 @@
 require 'serverspec'
+set :backend, :exec
 
 RSpec.configure do |c|
   c.path = '/usr/local/bin:/sbin:/bin:/usr/bin'